requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-amazon
dvc[s3]==3.50.1
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
boto3>=1.28.0,<1.34.0
botocore>=1.31.0,<1.34.0
gitpython==3.1.40
matplotlib==3.8.2
tensorboard==2.15.1
mlflow==2.8.1
scikit-learn==1.3.2